feat: add very basic png file viewing support
